modificando las validaciones

+++ Randy_Fanbian_P2_Ap1/Models/Vehiculo.cs
@@ -14,11 +14,11 @@ namespace Randy_Fanbian_P2_Ap1.Models
         
             [Required(ErrorMessage = "Este campo es requerido")]
             public string? Descripcion { get; set; }
-        
-            [Range(1, int.MaxValue, ErrorMessage = "El campo debe ser mayor que cero")]
+
+            [Range(1, (double)decimal.MaxValue, ErrorMessage = "El campo debe ser mayor que cero")]
             public decimal Costo { get; set; }
 
-            [Range(1, int.MaxValue, ErrorMessage = "El campo debe ser mayor que cero")]
+            [Range(1, (double)decimal.MaxValue, ErrorMessage = "El campo debe ser mayor que cero")]
             public decimal Gasto { get; set; }
 
             [ForeignKey("VehiculoId")]
